+++ rplatform/valid_notes2.R
@@ -1,7 +1,1 @@
-global_valid_notes_l <- list(
-  list(
-    length = 7,
-    index_to_check = 7,
-    text_to_check = "..i ..norm_method column"
-  )
-)
+global_valid_notes_l <- list()